0.2.0.0
Implemented a generic GameStateListener class to facilitate the calling
of functions that we only want to run once at during a particular game state.

Implemented a tiered gamestate/timer initialization system in order to 
(hopefully) properly initalize the script regardless of point of the point
at which it is invoked from the console (including pick phase).

Implemented the choosing of settings by team chat during PreGame.